@@ -4,38 +4,54 @@
   "private": true,
   "author": "whcraftmc",
   "scripts": {
-    "docusaurus": "docusaurus",
-    "start": "docusaurus start",
-    "build": "docusaurus build",
-    "swizzle": "docusaurus swizzle",
-    "deploy": "docusaurus deploy",
+    "dev": "docusaurus start --no-open --no-minify --config ./docusaurus.config.ts",
+    "start": "yarn build && yarn serve",
+    "build": "docusaurus build --config ./docusaurus.config.ts",
     "clear": "docusaurus clear",
-    "serve": "docusaurus serve",
+    "serve": "docusaurus serve --no-open --config ./docusaurus.config.ts",
+    "swizzle": "docusaurus swizzle --typescript",
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
-    "typecheck": "tsc"
+    "format": "prettier . --write"
   },
   "dependencies": {
-    "@docusaurus/core": "2.4.1",
-    "@docusaurus/preset-classic": "2.4.1",
-    "@mdx-js/react": "^1.6.22",
-    "clsx": "^1.2.1",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "vercel": "^31.1.1"
+    "@docusaurus/core": "0.0.0-5566",
+    "@docusaurus/plugin-content-docs": "0.0.0-5566",
+    "@docusaurus/plugin-content-pages": "0.0.0-5566",
+    "@docusaurus/plugin-debug": "0.0.0-5566",
+    "@docusaurus/plugin-pwa": "0.0.0-5566",
+    "@docusaurus/plugin-sitemap": "0.0.0-5566",
+    "@docusaurus/theme-classic": "0.0.0-5566",
+    "@docusaurus/theme-common": "0.0.0-5566",
+    "@docusaurus/theme-search-algolia": "0.0.0-5566",
+    "@fontsource/jetbrains-mono": "4.5.12",
+    "@mdx-js/react": "2.3.0",
+    "clsx": "1.2.1",
+    "prism-react-renderer": "2.0.3",
+    "react": "17.0.2",
+    "react-dom": "17.0.2"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.1",
-    "@docusaurus/tsconfig": "^0.0.0-5601",
-    "@tsconfig/docusaurus": "^1.0.5",
-    "typescript": "^4.9.5"
+    "@docusaurus/module-type-aliases": "0.0.0-5566",
+    "@docusaurus/types": "0.0.0-5566",
+    "@fec/remark-a11y-emoji": "3.1.0",
+    "@tsconfig/docusaurus": "1.0.7",
+    "@types/is-ci": "3.0.0",
+    "@types/node": "17.0.45",
+    "@types/prettier": "2.7.2",
+    "esbuild-loader": "3.0.1",
+    "esbuild-register": "3.4.2",
+    "is-ci": "3.0.1",
+    "prettier": "2.8.8",
+    "regenerator-runtime": "0.13.11",
+    "typescript": "5.0.2"
   },
   "browserslist": {
     "production": [
       ">0.5%",
       "not dead",
-      "not op_mini all"
+      "not op_mini all",
+      "not ie <= 11"
     ],
     "development": [
       "last 1 chrome version",
@@ -43,7 +59,9 @@
       "last 1 safari version"
     ]
   },
-  "engines": {
-    "node": ">=18"
+  "packageManager": "yarn@3.5.0",
+  "resolutions": {
+    "@docusaurus/core@0.0.0-5566": "patch:@docusaurus/core@npm%3A0.0.0-5566#./.yarn/patches/@docusaurus-core-npm-0.0.0-5566-187f2a3806.patch",
+    "@docusaurus/plugin-content-docs@0.0.0-5566": "patch:@docusaurus/plugin-content-docs@npm%3A0.0.0-5566#./.yarn/patches/@docusaurus-plugin-content-docs-npm-0.0.0-5566-e59592d84c.patch"
   }
 }
